backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic[email]==2.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pdfplumber==0.11.0
google-generativeai==0.7.2
python-dotenv==1.0.1